crsmeta: Extract Coordinate System Metadata

Obtain coordinate system metadata from various data formats. There are functions to extract a 'CRS' (coordinate reference system, <https://en.wikipedia.org/wiki/Spatial_reference_system>) in 'EPSG' (European Petroleum Survey Group, <http://www.epsg.org/>), 'PROJ4' <https://proj.org/>, or 'WKT2' (Well-Known Text 2, <http://docs.opengeospatial.org/is/12-063r5/12-063r5.html>) forms. This is purely for getting simple metadata from in-memory formats, please use other tools for out of memory data sources.

Obtain coordinate system metadata from various in-memory R formats.

Details

Coordinate reference systems (CRS) are described at https://en.wikipedia.org/wiki/Spatial_reference_system. The European Petroleum Survey Group (EPSG) codes are maintained at http://www.epsg.org/. The old 'PROJ4' forms (along with newer systems) are described with the 'PROJ' library https://proj.org/. The Well-Known Text system 'WKT2' is documented at http://docs.opengeospatial.org/is/12-063r5/12-063r5.html.

There are a few functions to return a character string of length one of each main kind of projection string.

crs_epsg	short numeric EPSG lookup

crs_proj	short PROJ4 style string

crs_wkt2	the full parameter set as WKT2 string

See also