Extract Coordinate System Metadata

Obtain coordinate system metadata from various in-memory R formats.

Details

Coordinate reference systems (CRS) are described at https://en.wikipedia.org/wiki/Spatial_reference_system. The European Petroleum Survey Group (EPSG) codes are maintained at http://www.epsg.org/. The old 'PROJ4' forms (along with newer systems) are described with the 'PROJ' library https://proj.org/. The Well-Known Text system 'WKT2' is documented at http://docs.opengeospatial.org/is/12-063r5/12-063r5.html.

There are a few functions to return a character string of length one of each main kind of projection string.

crs_epsg	short numeric EPSG lookup

crs_proj	short PROJ4 style string

crs_wkt2	the full parameter set as WKT2 string

Note: this reference page has a strange name 'crsmeta1-package' because for some reason 'crsmeta-package' does not show up in pkgdown output.